Day 5 submission (sorry i didn't finish part 2 :/)
