--- Day5/VSProject/x64/Debug/Day5.tlog/Day5.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.32.31326:TargetPlatformVersion=10.0.19041.0:
-Debug|x64|D:\repos\week4_austinmarckx\Day5\|
