+++ Day5_pt2/CMakeLists.txt
@@ -1,0 +1,4 @@
+project(Day5)
+
+add_executable(fm
+        src/main.cpp)
