+++ Day5_pt2/Day5.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.32413.511
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Day5", "VSProject\Day5.vcxproj", "{F14729E1-D2BF-46CD-962F-AB23B450F1E6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F14729E1-D2BF-46CD-962F-AB23B450F1E6}.Debug|x64.ActiveCfg = Debug|x64
+		{F14729E1-D2BF-46CD-962F-AB23B450F1E6}.Debug|x64.Build.0 = Debug|x64
+		{F14729E1-D2BF-46CD-962F-AB23B450F1E6}.Debug|x86.ActiveCfg = Debug|Win32
+		{F14729E1-D2BF-46CD-962F-AB23B450F1E6}.Debug|x86.Build.0 = Debug|Win32
+		{F14729E1-D2BF-46CD-962F-AB23B450F1E6}.Release|x64.ActiveCfg = Release|x64
+		{F14729E1-D2BF-46CD-962F-AB23B450F1E6}.Release|x64.Build.0 = Release|x64
+		{F14729E1-D2BF-46CD-962F-AB23B450F1E6}.Release|x86.ActiveCfg = Release|Win32
+		{F14729E1-D2BF-46CD-962F-AB23B450F1E6}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {298523BF-9564-4685-B592-643111457851}
+	EndGlobalSection
+EndGlobal
